+++ docs/HISTORY.rst
@@ -7,6 +7,7 @@ Changelog
 - Fixed links for title [kaerumy]
 - Added keyword field to listing, and form [kaerumy]
 - Add download csv feature to goodpracticelisting [kaerumy]
+- Fix handling for unicode and date fields [kaerumy]
 
 
 1.0 (2014-01-29)